lambda/requirements.txt
# Bedrock QA Validation System - Lambda Dependencies

# Strands Agent Framework
strands-agents>=0.1.0
strands-agents-tools>=0.1.0

# AWS SDK
boto3>=1.34.0
botocore>=1.34.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
hypothesis>=6.92.0
moto>=5.0.0

# Utilities
python-dateutil>=2.8.2
